chore: Adjust StorageFile.GetFileFromApplicationUriAsync for iOS/Android
