+++ src/Uno.UWP/Helpers/AndroidResourceNameEncoder.cs
@@ -47,6 +47,7 @@ namespace Uno
 		}
 
 		public static string EncodeFileSystemPath(string path)
+			// Android assets need to placed in the Assets folder
 			=> global::System.IO.Path.Combine("Assets", EncodePath(path, global::System.IO.Path.DirectorySeparatorChar));
 
 		public static string EncodeResourcePath(string path)
