+++ src/Uno.UWP/Storage/StorageFile.android.cs
@@ -10,16 +10,20 @@ using System.Threading.Tasks;
 using Android.Content.Res;
 using Uno;
 using Uno.Extensions;
+using Uno.Logging;
 using Uno.UI;
 using Windows.Foundation;
 using Windows.Storage.FileProperties;
 using Windows.Storage.Streams;
 using Windows.UI.Composition.Interactions;
+using Windows.Storage.Helpers;
 
 namespace Windows.Storage
 {
 	public partial class StorageFile : StorageItem, IStorageFile
 	{
+		private static ConcurrentEntryManager _assetGate = new ConcurrentEntryManager();
+
 		private static async Task<StorageFile> GetFileFromApplicationUriAsyncTask(CancellationToken ct, Uri uri)
 		{
 			if(uri.Scheme != "ms-appx")
@@ -33,10 +37,26 @@ namespace Windows.Storage
 			var assets = global::Android.App.Application.Context.Assets;
 			var outputCachePath = global::System.IO.Path.Combine(Android.App.Application.Context.CacheDir.AbsolutePath, path);
 
-			using var input = assets.Open(path);
+			if (typeof(StorageFile).Log().IsEnabled(Microsoft.Extensions.Logging.LogLevel.Debug))
+			{
+				typeof(StorageFile).Log().Debug($"GetFileFromApplicationUriAsyncTask path:{path} outputCachePath:{outputCachePath}");
+			}
 
-			using var output = File.OpenWrite(outputCachePath);
-			await input.CopyToAsync(output);
+			// Ensure only one generation of the cached file can occur at a time.
+			// The copy of tha android asset should not be necessary, but requires a significant
+			// refactoring of the StorageFile class to make non-local provides opaque.
+			// This will need to be revisited once we add support for other providers.
+			using var gate = await _assetGate.LockForAsset(ct, path);
+
+			if (!File.Exists(outputCachePath))
+			{
+				Directory.CreateDirectory(global::System.IO.Path.GetDirectoryName(outputCachePath));
+
+				using var input = assets.Open(path);
+				using var output = File.OpenWrite(outputCachePath);
+
+				await input.CopyToAsync(output);
+			}
 
 			return await StorageFile.GetFileFromPathAsync(outputCachePath);
 		}
